Feature/reentrancy (#111)
* gossip confirmance fixes #1

* gossip: bugfixes

* gossip: more fixes

* gossip: build related minor changes

* gossip: one more fix

* gossip: flush hello message asap

* experiment with noise protocol

* made injector creation fns inline to prevent multiple definitions

* gossip restructured, interop issues fixed

* Noise::write now reports the correct amount of bytes


* refactoring: move prev implementation of Kademlia to other namespace


* feature: kademlia dependency injection


* feature: improve random generator


* fix: getting listened multiaddress in tcp listener


* fix: reduce logging in secio


* draft: new implementation of kademlia


* draft: processing with kademlia


* draft: continue processing with kademlia


* draft: continue


* draft: kademlia


* feature: validation


* fix: DSA


* feature: CIDv1 encoding
refactoring: replace deprecated sha256 function by hasher


* fix: multiaddress operations


* feature: make PeerId comparable to using in std::set


* refactoring: addr repo


* fix: Go-implementation compatibility


* feature: rendezvous chat as example of Kademlia using


* feature: unit-test for kademlia parts


* fix: StorageBackend interface
fix: ContentValue


* wipe: remove previous implementation of Kademlia


* fix: cmake files


cmake


* fix: some warnings


* fix: resolve TODOes, comments, format


* fix: clang-tidy issues


* fix: clang-tidy issues


* fix: clang-tidy issues


* feature: using timeout for make new stream
feature: smart using of peer routing


* fix: kademlia message parsing


* refactoring: optimize executors' working


* refactoring: optimize start of rendezvous chat


* refactoring: headers including


* refactoring: remaining request executors; naming


* feature: add handle and timeout as argument of Host::connect


* fix: some log messages


* injectors temp fix

* gossip: uncommented writing bytes checking

* removed redundant std::hash definition

* Hack yamux to allow weighty messages processing


* fix vtable

* fixes in DI injectors

* yamux test corrected according to curent window update policy

* added buffering primitives

* defer* functions in Reader/Writer interfaces and Yamux redesign pt.1

* scheduler fix regarding move assignment + cancel

* write queue interface change

* read buffer fix regarding subspan

* some diagnostic logging

* tcp connection fixes related to closing behavior

* yamux bugfixes

* yamux tests regression WIP

* build fix

* .

* .

* bugfixes

* fixes

* .

* .

* suppressed most verbose logging in yamux and multiselect

* yamux stream adjustWindowSize adjusted

* fixes regarding std::move of r/w callbacks (against possible ptrs invalidation)

* test cases with jumbo messages transfer added for yamux on noise/tls/plaintext

* bugfixes related to yamux window sizes, overflow, and acknowledgements

* echo protocol and examples support very large msgs

* changes in tests

* all muxers acceptance test recovered

* all muxers acceptance test: fixed issue with mock lifetime

* CI fixes

* yamux refactorings helped to avoid memory issues caused by reentrant callbacks

* build fix for mac

* another yamux fix to avoid memory issues caused by reentrant callbacks

* gossip: hotfix related to unbanning peers

* minor fixes reflecting review feedback

* temp loggers workaround

* Feature/multiselect upd (#121)

* multiselect revised, WIP

* multiselect: simple outbound stream negotiate

* multiselect numerous fixes

* multiselect: instances and reuse

* multiselect: fixes

* multiselect: removed old implementation

* multiselect: interop with go impl fixes

* multiselect: bugfixes

* multiselect: ProtocolMuxer interface abstracts simple outbound stream negotiation

* multiselect: cleanups and logging

* trigger CI

* temporarily disabled tests that required synchronous reaction of multiselect

* just removed unused lines

* reverted back ci.yml

+++ src/security/noise/crypto/state.cpp
@@ -44,7 +44,7 @@ namespace libp2p::security::noise {
 
   outcome::result<Key32> bytesToKey32(gsl::span<const uint8_t> key) {
     Key32 result;
-    if (key.size() != result.size()) {
+    if (static_cast<size_t>(key.size()) != result.size()) {
       return Error::WRONG_KEY32_SIZE;
     }
     std::copy_n(key.begin(), result.size(), result.begin());
